fix(admin-ui): move test helpers to fixtures, pass file size gate
admin.spec.ts (3227 lines) exceeded the 3000-line frontend limit.
Move helper functions (chooseSidebarSelectOption, openIntegrationEditor,
disableTestMotion) and the mockAdminApi route mock to admin.fixtures.ts.
The spec file drops to ~1663 lines, well under the limit.

Fixes: File size gate failure on PR #1983
